Implement tile-to-tile walking

+++ src/component/camera.rs
@@ -1,4 +1,4 @@
 pub struct Camera {
-  pub view_width: f32,
-  pub view_height: f32
+  pub view_width: f64,
+  pub view_height: f64
 }
